Some small corrections
Hi Jürgen,
very inspiring article. I changed 2 misspellings and I removed a lot of colons - maybe I'm wrong here but I think the English  language needs colons only for series like I did this, that an d that. Not for every secondary sentence as in German.
But feel free to decide yourself :)